Refactor renderers
