+++ eps/eps.go
@@ -42,19 +42,27 @@ xrad yrad scale
 savematrix setmatrix
 } def`
 
-type Renderer struct {
+// Writer writes the canvas as an EPS file.
+// Be aware that EPS does not support transparency of colors.
+func Writer(w io.Writer, c *canvas.Canvas) error {
+	eps := New(w, c.W, c.H)
+	c.Render(eps)
+	return nil
+}
+
+type EPS struct {
 	w             io.Writer
 	width, height float64
 	color         color.RGBA
 }
 
 // New creates an encapsulated PostScript renderer.
-func New(w io.Writer, width, height float64) *Renderer {
+func New(w io.Writer, width, height float64) *EPS {
 	fmt.Fprintf(w, "%%!PS-Adobe-3.0 EPSF-3.0\n%%%%BoundingBox: 0 0 %v %v\n", dec(width), dec(height))
 	fmt.Fprintf(w, psEllipseDef)
 	// TODO: (EPS) generate and add preview
 
-	return &Renderer{
+	return &EPS{
 		w:      w,
 		width:  width,
 		height: height,
@@ -62,18 +70,18 @@ func New(w io.Writer, width, height float64) *Renderer {
 	}
 }
 
-func (r *Renderer) setColor(color color.RGBA) {
+func (r *EPS) setColor(color color.RGBA) {
 	if color != r.color {
 		fmt.Fprintf(r.w, " %v %v %v setrgbcolor", dec(float64(color.R)/255.0), dec(float64(color.G)/255.0), dec(float64(color.B)/255.0))
 		r.color = color
 	}
 }
 
-func (r *Renderer) Size() (float64, float64) {
+func (r *EPS) Size() (float64, float64) {
 	return r.width, r.height
 }
 
-func (r *Renderer) RenderPath(path *canvas.Path, style canvas.Style, m canvas.Matrix) {
+func (r *EPS) RenderPath(path *canvas.Path, style canvas.Style, m canvas.Matrix) {
 	// TODO: (EPS) test ellipse, rotations etc
 	// TODO: (EPS) add drawState support
 	// TODO: (EPS) use dither to fake transparency
@@ -83,12 +91,12 @@ func (r *Renderer) RenderPath(path *canvas.Path, style canvas.Style, m canvas.Ma
 	r.w.Write([]byte(" fill"))
 }
 
-func (r *Renderer) RenderText(text *canvas.Text, m canvas.Matrix) {
+func (r *EPS) RenderText(text *canvas.Text, m canvas.Matrix) {
 	// TODO: (EPS) write text natively
 	text.RenderAsPath(r, m)
 }
 
-func (r *Renderer) RenderImage(img image.Image, m canvas.Matrix) {
+func (r *EPS) RenderImage(img image.Image, m canvas.Matrix) {
 	// TODO: (EPS) write image
 }
