+++ svg/svg_test.go
